calculate organism, arrange biosample interval properties
