+++ src/encoded/schemas/donor.json
@@ -24,14 +24,6 @@
                 "pattern": "^(ENCDO\\d{3}[A-Z]{3})$|^(TSTDO\\d{6})$"
             }
         },
-        "organism": {
-            "title": "Organism",
-            "description": "Common name of donor organism.",
-            "comment": "Do not submit, value is assigned by the object.",
-            "permission": "import_items",
-            "type": "string",
-            "linkTo": "Organism"
-        },
         "treatments": {
             "description": "A list of treatments applied to the donor.",
             "comment": "Can include drug treatments, medication, and/or conditions like fasting"
