+++ src/encoded/schemas/human_prenatal_donor.json
@@ -4,7 +4,7 @@
     "id": "/profiles/human_prenatal_donor.json",
     "$schema": "http://json-schema.org/draft-04/schema#",
     "type": "object",
-    "required": ["organism", "life_stage", "sex"],
+    "required": ["life_stage", "sex"],
     "identifyingProperties": ["uuid", "accession", "aliases"],
     "additionalProperties": false,
     "mixinProperties": [
