+++ src/encoded/tests/data/inserts/mouse_postnatal_donor.json
@@ -1,7 +1,6 @@
 [
     {
         "accession": "ENCDO012AAA",
-        "organism": "mouse",
         "strain_background": "CD-1",
         "documents": [
             "lattice:PanIsletD_Crawford_protocol"
@@ -13,7 +12,6 @@
     },
     {
         "accession": "ENCDO015AAA",
-        "organism": "mouse",
         "strain_background": "129/Ola",
         "uuid": "1a8dd515-63da-40d8-a166-351858c77fd7",
         "life_stage": "adult",
@@ -22,7 +20,6 @@
     },
     {
         "accession": "ENCDO001MUS",
-        "organism": "mouse",
         "strain_background": "unknown",
         "uuid": "f08acb61-5af6-4b7d-a7c8-f08be52d8012",
         "life_stage": "adult",
